@@ -11,23 +11,23 @@
   "dependencies": {
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
-    "@mui/icons-material": "^5.10.16",
-    "@mui/material": "^5.10.17",
-    "@mui/styled-engine-sc": "^5.10.16",
-    "@next/font": "^13.0.6",
-    "@react-spring/web": "^9.6.0",
-    "next": "13.0.6",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.11.3",
+    "@mui/styled-engine-sc": "^5.11.0",
+    "@next/font": "^13.1.1",
+    "@react-spring/web": "^9.6.1",
+    "next": "13.1.1",
     "node-sass": "^7.0.3",
-    "query-string": "^8.0.0",
+    "query-string": "^8.1.0",
     "react": "18.2.0",
     "react-auth-code-input": "^3.2.1",
     "react-dom": "18.2.0",
-    "sass": "^1.56.2",
+    "sass": "^1.57.1",
     "styled-components": "^5.3.6",
     "typescript": "^4.9.4"
   },
   "devDependencies": {
-    "eslint": "8.29.0",
-    "eslint-config-next": "13.0.6"
+    "eslint": "8.31.0",
+    "eslint-config-next": "13.1.1"
   }
 }
